requirements.txt
python-telegram-bot==21.*
scikit-learn
